Fix CI regressions on inheritdoc PR
Revert accidental SDK/MSBuild SDK updates and restore nullable warning settings for net462. Also split InheritdocShouldNotBeUsedOnTypesAnalyzerTests into a matching file to satisfy MA0048 and keep the new cref test coverage.
