+++ Directory.Build.targets
@@ -100,7 +100,7 @@
   </Choose>
 
   <!-- disable the nullable warnings when compiling for .NET Standard 2.0 -->
-  <PropertyGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+  <PropertyGroup Condition="'$(TargetFramework)' == 'netstandard2.0' OR '$(TargetFramework)' == 'net462'">
     <NoWarn>$(NoWarn);nullable</NoWarn>
   </PropertyGroup>
 </Project>
