+++ global.json
@@ -1,11 +1,11 @@
 {
   "sdk": {
-    "version": "10.0.203",
+    "version": "10.0.100",
     "rollForward": "latestPatch"
   },
   "msbuild-sdks": {
-    "Meziantou.NET.Sdk": "1.0.92",
-    "Meziantou.NET.Sdk.Test": "1.0.92",
-    "Meziantou.NET.Sdk.Web": "1.0.92"
+    "Meziantou.NET.Sdk": "1.0.42",
+    "Meziantou.NET.Sdk.Test": "1.0.42",
+    "Meziantou.NET.Sdk.Web": "1.0.42"
   }
 }
